+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 * [#6161](https://github.com/rubocop-hq/rubocop/pull/6161): Add scope methods to `Rails/FindEach` cop. Makes the cop also check for the following scopes: `eager_load`, `includes`, `joins`, `left_joins`, `left_outer_joins`, `preload`, `references`, and `unscoped`. ([@repinel][])
 * [#6137](https://github.com/rubocop-hq/rubocop/pull/6137): Allow `db` to allowed names of `Naming/UncommunicativeMethodParamName` cop in default config. ([@mkenyon][])
+* Update the highlighting of `Lint/DuplicateMethods` to include the method name. ([@rrosenblum][])
 
 ## 0.58.2 (2018-07-23)